test: merge loaded-model-info tests into model-info family
