+++ packages/sdk/tests-qvac/tests/desktop/consumer.ts
@@ -58,7 +58,6 @@ import { ConfigReloadExecutor } from "./executors/config-reload-executor.js";
 import { LoggingExecutor } from "../shared/executors/logging-executor.js";
 import { RegistryExecutor } from "../shared/executors/registry-executor.js";
 import { ModelInfoExecutor } from "../shared/executors/model-info-executor.js";
-import { LoadedModelInfoExecutor } from "../shared/executors/loaded-model-info-executor.js";
 import { WrongModelExecutor } from "../shared/executors/wrong-model-executor.js";
 import { ErrorExecutor } from "../shared/executors/error-executor.js";
 import { TtsExecutor } from "../shared/executors/tts-executor.js";
@@ -349,7 +348,6 @@ export const executor = createExecutor({
     new EmbeddingExecutor(resources),
     new RagExecutor(resources),
     new ModelInfoExecutor(resources),
-    new LoadedModelInfoExecutor(resources),
     new WrongModelExecutor(resources),
     new ErrorExecutor(resources),
     new ToolsExecutor(resources),
